pyproject.toml
[project]
name = "swmgemini"
version = "0.1.0"
description = "Hackathon Project"
readme = "README.md"
requires-python = "==3.11.*"
dependencies = [
    "fastapi>=0.135.2",
    "fishjam-server-sdk[gemini]>=0.25.0",
    "google-cloud-texttospeech>=2.0.0",
    "google-genai>=1.0.0",
    "pillow>=12.1.1",
    "pydub>=0.25.1",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu126"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
